requirements.txt
gensim==3.8.1
numpy>=1.21
regex==2019.12.20
scipy==1.3.2
sentencepiece==0.1.85
sklearn==0.0
spacy==3.1
torch==1.9.0
transformers==4.9.2
tqdm==4.41.0
jsonlines==1.2.0
nltk==3.5
pandas
pingouin
openai
tenacity
python-dotenv
crowd-kit
googletrans
